workspacesView: Add an explicit show() method
Currently we rely on Overview::showing to trigger the overview
transition. However with recent code changes, the WorkspacesView
objects themselves are created in response to that signal, and
given that they are destroyed on Overview::hide, the signal
handler that handles the transition is now never called. To fix,
make the handler a public function and call it from WorkspacesDisplay
as necessary.

Signed-off-by: Joost Verdoorn <jpverdoorn@gmail.com>